+++ app/Info.php
@@ -19,7 +19,7 @@ class Info extends Model
     public function save(array $options = [])
     {
         // Empty features, we're not saving anything
-        if (empty($this->attributes['features'])) return;
+        if (is_array($this->attributes['features']) && empty($this->attributes['features'])) return;
 
         try {
             unset($this->identities);