+++ database/migrations/20200514143915_change_picture_type_messages_table.php
@@ -1,0 +1,30 @@
+<?php
+
+use Movim\Migration;
+use Illuminate\Database\Schema\Blueprint;
+
+class ChangePictureTypeMessagesTable extends Migration
+{
+    public function up()
+    {
+        $this->schema->table('messages', function (Blueprint $table) {
+            $table->dropColumn('picture');
+        });
+
+        $this->schema->table('messages', function (Blueprint $table) {
+            $table->boolean('picture')->default(false);
+            $table->index('picture');
+        });
+    }
+
+    public function down()
+    {
+        $this->schema->table('messages', function (Blueprint $table) {
+            $table->dropColumn('picture');
+        });
+
+        $this->schema->table('messages', function (Blueprint $table) {
+            $table->text('picture')->nullable();
+        });
+    }
+}