+++ CHANGELOG.md
@@ -24,6 +24,7 @@ v0.18 (trunk)
  * Add support for Unicode 12.0 emojis
  * Add a refresh system for the Chat header based on the presences (and filtered by notifications)
  * Use UNION ALL instead of OR for messages request (to prevent optimisations issues in SQL)
+ * Better handling of Pictures in Posts and Messages
 
 v0.17.1
 ---------------------------